feat: version 2 catalog serialization
Changed from using bitcode for the catalog to JSON. This applies to both
the log files as well as snapshots.

This required copying existing code into two places:
- v1
- latest

The types from latest are used as the "in-memory" types that we can work
with throughout the codebase. The v1 types are only there for posterity
and giving the ability to deserialize catalog files that predate this
change; they are not meant to be used throughout the code.
