+++ influxdb3_catalog/src/log/versions/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod latest;
+pub(crate) mod v1;
